+++ backend/app/services/llm/mappers.py
@@ -99,6 +99,10 @@ def map_kaapi_to_google_params(kaapi_params: dict) -> tuple[dict, list[str]]:
     warnings = []
 
     # Model is present in all param types
+    model = kaapi_params.get("model")
+    if not model:
+        return {}, ["Missing required 'model' parameter"]
+
     google_params["model"] = kaapi_params.get("model")
 
     # Instructions for STT prompts
